ShopBot-AI-Bot/requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
transformers
torch
openpyxl
xlrd
sentencepiece
